requirements.txt
torch~=1.9.1+cu111
opencv-python~=4.11.0.86
onnxruntime~=1.19.2
setuptools~=75.8.0
numpy~=1.26.0
matplotlib~=3.9.4
pandas~=2.0.1
pydicom~=2.4.4
simpleitk~=2.4.1
nibabel~=5.3.2
pillow~=11.1.0
torchvision~=0.10.1+cu111